requirements.txt
absl-py==1.0.0
altair==4.1.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==21.4b2
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
detectron2==0.5+cu111
entrypoints==0.4
future==0.18.2
fvcore==0.1.5.post20220212
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
iopath==0.1.8
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.2
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.19.5
oauthlib==3.2.0
omegaconf==2.1.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
portalocker==2.4.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pydeck==0.6.2
pydot==1.4.2
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==0.24.2
scipy==1.5.4
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
streamlit==1.8.1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.11.2
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tornado==6.1
tqdm==4.63.0
traitlets==4.3.3
typed-ast==1.5.2
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
yacs==0.1.8
zipp==3.6.0
absl-py==1.0.0
altair==4.1.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==21.4.0
augly==1.0.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==21.4b2
bleach==4.1.0
blinker==1.4
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
detectron2==0.5+cu111
entrypoints==0.4
future==0.18.2
fvcore==0.1.5.post20220212
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
h5py==3.1.0
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
iopath==0.1.8
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.2
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.19.5
oauthlib==3.2.0
omegaconf==2.1.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
portalocker==2.4.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pydeck==0.6.2
pydot==1.4.2
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-magic==0.4.25
pytorch-metric-learning==1.3.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==0.24.2
scipy==1.5.4
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
streamlit==1.8.1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
timm==0.5.4
toml==0.10.2
toolz==0.11.2
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tornado==6.1
tqdm==4.63.0
traitlets==4.3.3
typed-ast==1.5.2
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
yacs==0.1.8
zipp==3.6.0
